+++ maestro-client/src/main/java/maestro/device/Platform.kt
@@ -3,13 +3,16 @@ package maestro.device
 enum class Platform(val description: String) {
     ANDROID("Android"),
     IOS("iOS"),
-    WEB("Web");
+    WEB("Web"),
+    DESKTOP("Desktop");
 
     companion object {
         fun fromString(p: String): Platform {
-            return entries.firstOrNull { it.description.equals(p, ignoreCase = true) }
+            return entries.firstOrNull {
+                it.description.equals(p, ignoreCase = true) || it.name.equals(p, ignoreCase = true)
+            }
                 ?: throw IllegalArgumentException(
-                    "Unknown platform: '$p'. Must be one of: ${entries.joinToString { it.description }}"
+                    "Unknown platform: '$p'. Must be one of: ${entries.joinToString { it.description }}",
                 )
         }
     }