+++ maestro-desktop/src/main/kotlin/maestro/drivers/desktop/DesktopDriverFactory.kt
@@ -1,0 +1,16 @@
+package maestro.drivers.desktop
+
+import maestro.Driver
+import maestro.drivers.desktop.linux.LinuxDesktopDriver
+import maestro.drivers.desktop.macos.MacOSDesktopDriver
+import maestro.drivers.desktop.windows.WindowsDesktopDriver
+
+object DesktopDriverFactory {
+    fun create(): Driver {
+        return when (DesktopOs.current()) {
+            DesktopOs.MACOS -> MacOSDesktopDriver()
+            DesktopOs.WINDOWS -> WindowsDesktopDriver()
+            DesktopOs.LINUX -> LinuxDesktopDriver()
+        }
+    }
+}